refactor(views): extract lyrics rendering from aside
Move lyric layout, Unicode wrapping, karaoke reveal, and paint helpers out
of the sidebar coordinator into private lyrics modules.

Keep Aside responsible for entity state, queue rendering, viewport motion,
scroll/follow behavior, and top-level composition.

The extraction preserves all rendering, timing, wrapping, caching, fade,
and spring behavior. All 38 existing aside tests remain unchanged, with
28 lyric-specific tests moved alongside the extracted implementation.
