+++ crates/views/src/chrome/aside.rs
@@ -1,33 +1,33 @@
 use std::{collections::HashMap, ops::Range};
-use unicode_segmentation::UnicodeSegmentation;
 
 use gpui::prelude::*;
-use icu_segmenter::{LineSegmenter, options::LineBreakOptions};
 
 use gpui::{
-    Animation, AnimationExt as _, App, Bounds, ContentMask, Context, DecorationRun, Div,
-    DragMoveEvent, Element, Entity, FontWeight, GlobalElementId, Hsla, InspectorElementId,
-    LayoutId, LineLayout, MouseDownEvent, Pixels, Point, Render, ScrollHandle, ScrollStrategy,
-    ScrollWheelEvent, ShapedLine, SharedString, SpringConfig, SpringState, Style, Task, TextAlign,
-    UniformListScrollHandle, Window, div, ease_in_out, point, px, svg, uniform_list,
+    Animation, AnimationExt as _, App, Bounds, Context, Div, DragMoveEvent, Entity, FontWeight,
+    MouseDownEvent, Pixels, Point, Render, ScrollHandle, ScrollStrategy, ScrollWheelEvent,
+    SharedString, SpringConfig, SpringState, Task, UniformListScrollHandle, Window, div,
+    ease_in_out, px, svg, uniform_list,
 };
 use i18n::t;
-use music::{Track, Voice};
+use music::Track;
 use router::{Destination, LibraryTab, Link as _, LocalTab};
 use state::{
-    AppSettings, Lyrics, LyricsState, Playback, PlaybackState, Queue, RomanizationScripts, SideTab,
-    Sonora, Whence,
+    AppSettings, Lyrics, LyricsState, Playback, PlaybackState, Queue, SideTab, Sonora, Whence,
 };
 use ui::{
-    ActiveTheme as _, Button, Card, DraggedPin, Edge, Motion, Motioned as _, Pin, Pinnable as _,
-    Popup, Scrollbar, Scroller, Spot, Springs, Text, Vacancy, drop_gap, drop_marker, ease_out_expo,
-    eyebrow, faint, mix, snapped, vacant,
+    ActiveTheme as _, Button, Card, DraggedPin, Edge, Motion, Pin, Pinnable as _, Popup, Scrollbar,
+    Scroller, Spot, Springs, Text, Vacancy, drop_gap, drop_marker, ease_out_expo, eyebrow, faint,
+    mix, snapped, vacant,
 };
 
 use crate::chrome::{Chrome, section_label};
 use crate::shared::effects;
 use crate::shared::menus::ItemMenu;
 use crate::shared::pins::Pinned as _;
+
+mod lyrics;
+
+use lyrics::*;
 
 const QUEUE: &str = "queue";
 const BULLET: SharedString = SharedString::new_static("·");
@@ -39,8 +39,6 @@ const BACKGROUND_SINGING_BLUR: Pixels = px(0.75);
 const VEIL: f32 = 0.3;
 const HAZE: f32 = 0.45;
 const VERSE_FADE: f32 = 1.25;
-const PAST: f32 = 0.4;
-const AHEAD: f32 = 0.6;
 const ACTIVE_VERSE_GROWTH: Pixels = px(2.);
 const FULLSCREEN_VERSE_GROWTH: Pixels = px(3.);
 const LYRICS_HORIZONTAL_INSET_REM: f32 = 1.5;
@@ -70,14 +68,6 @@ const RESOLVE_BLUR: f32 = 0.2;
 const RESOLVE_FADE: f32 = 0.5;
 const SETTLE: std::time::Duration = std::time::Duration::from_secs(4);
 const INSTRUMENTAL_BREAK: std::time::Duration = std::time::Duration::from_secs(5);
-const SWEEP_LEAST: std::time::Duration = std::time::Duration::from_millis(180);
-// karaoke sweep ceiling
-const KARAOKE_HZ: u32 = 45;
-const KARAOKE_FRAME: std::time::Duration =
-    std::time::Duration::from_nanos(1_000_000_000 / KARAOKE_HZ as u64);
-// what a lane row actually takes, plus the gaps between lanes
-const LANE_GAP_REM: f32 = 0.25;
-const LANE_SLACK: f32 = 0.25;
 
 fn track(queue: &Queue, position: QueuePosition) -> Option<Track> {
     match position {
@@ -183,17 +173,6 @@ struct Touch {
     depth: f32,
     waking: bool,
     settling: bool,
-}
-
-#[derive(Clone, Copy)]
-struct Sung {
-    karaoke: bool,
-    lane: Pixels,
-    scripts: Option<RomanizationScripts>,
-    theme: ui::Theme,
-    karaoke_tint: gpui::Hsla,
-    lift: f32,
-    from: gpui::Point<f32>,
 }
 
 #[derive(Clone, Copy)]
@@ -1866,396 +1845,6 @@ fn source_link(name: SharedString, to: Destination, cx: &App) -> impl IntoElemen
         .child(name)
 }
 
-fn fixed_lyrics_lane(rows: &[SharedString], voice: Voice, sung: Sung) -> Div {
-    div()
-        .flex()
-        .flex_col()
-        .children(rows.iter().map(move |row| {
-            lifted(
-                div()
-                    .w_full()
-                    .when(!voice.lead(), |this| this.text_right())
-                    .child(row.clone()),
-                sung,
-            )
-        }))
-}
-
-fn karaoke_lane(
-    plan: &Wrapped,
-    words: &[music::LyricsWord],
-    position: std::time::Duration,
-    voice: Voice,
-    sung: Sung,
-) -> Div {
-    let windows = (0..words.len())
-        .map(|word| karaoke_window(words, word))
-        .collect::<Vec<_>>();
-    let lit = |shape: ShapedLine, reveal: Vec<RevealRange>| {
-        KaraokeText::new(shape, reveal, sung.karaoke_tint)
-    };
-
-    match plan.rows.is_empty() {
-        false => div()
-            .flex()
-            .flex_col()
-            .text_left()
-            .children((0..plan.rows.len()).map(|row| {
-                let reveal = revealed(plan, plan.rows[row].clone(), &windows, position);
-                lifted(
-                    div()
-                        .flex()
-                        .when(!voice.lead(), |this| this.justify_end())
-                        .child(lit(plan.shapes[row].clone(), reveal)),
-                    sung,
-                )
-            })),
-        true => div()
-            .flex()
-            .flex_wrap()
-            .text_left()
-            .when(!voice.lead(), |this| this.justify_end())
-            .children((0..plan.units.len()).map(|index| {
-                let reveal = revealed(plan, index..index + 1, &windows, position);
-                lit(plan.shapes[index].clone(), reveal)
-            })),
-    }
-}
-
-struct KaraokeText {
-    line: ShapedLine,
-    ranges: Vec<RevealRange>,
-    foreground: Hsla,
-}
-
-struct KaraokeTextLayout {
-    line_height: Pixels,
-    base: DecorationRun,
-    foreground: DecorationRun,
-}
-
-impl KaraokeText {
-    fn new(line: ShapedLine, ranges: Vec<RevealRange>, foreground: Hsla) -> Self {
-        Self {
-            line,
-            ranges,
-            foreground,
-        }
-    }
-}
-
-impl IntoElement for KaraokeText {
-    type Element = Self;
-
-    fn into_element(self) -> Self::Element {
-        self
-    }
-}
-
-impl Element for KaraokeText {
-    type RequestLayoutState = KaraokeTextLayout;
-    type PrepaintState = ();
-
-    fn id(&self) -> Option<gpui::ElementId> {
-        None
-    }
-
-    fn source_location(&self) -> Option<&'static core::panic::Location<'static>> {
-        None
-    }
-
-    fn request_layout(
-        &mut self,
-        _: Option<&GlobalElementId>,
-        _: Option<&InspectorElementId>,
-        window: &mut Window,
-        cx: &mut App,
-    ) -> (LayoutId, Self::RequestLayoutState) {
-        let style = window.text_style();
-        let base_run = style.to_run(self.line.len());
-        let line_height = style.line_height_in_pixels(window.rem_size());
-
-        let mut layout_style = Style {
-            flex_shrink: 0.,
-            ..Style::default()
-        };
-        layout_style.size.width = self.line.width().into();
-        layout_style.size.height = line_height.into();
-
-        let base = DecorationRun {
-            len: self.line.len() as u32,
-            color: base_run.color,
-            background_color: base_run.background_color,
-            underline: base_run.underline,
-            strikethrough: base_run.strikethrough,
-        };
-        let foreground = DecorationRun {
-            len: self.line.len() as u32,
-            color: self.foreground,
-            background_color: base_run.background_color,
-            underline: base_run.underline,
-            strikethrough: base_run.strikethrough,
-        };
-
-        (
-            window.request_layout(layout_style, [], cx),
-            KaraokeTextLayout {
-                line_height,
-                base,
-                foreground,
-            },
-        )
-    }
-
-    fn prepaint(
-        &mut self,
-        _: Option<&GlobalElementId>,
-        _: Option<&InspectorElementId>,
-        _: Bounds<Pixels>,
-        _: &mut Self::RequestLayoutState,
-        _: &mut Window,
-        _: &mut App,
-    ) -> Self::PrepaintState {
-    }
-
-    fn paint(
-        &mut self,
-        _: Option<&GlobalElementId>,
-        _: Option<&InspectorElementId>,
-        bounds: Bounds<Pixels>,
-        layout: &mut Self::RequestLayoutState,
-        _: &mut Self::PrepaintState,
-        window: &mut Window,
-        cx: &mut App,
-    ) {
-        LineLayout::paint(
-            &self.line,
-            bounds.origin,
-            layout.line_height,
-            TextAlign::Left,
-            None,
-            std::slice::from_ref(&layout.base),
-            window,
-            cx,
-        )
-        .ok();
-
-        let parent_mask = window.content_mask();
-        let line_width = self.line.width();
-        for range in self.ranges.iter().copied() {
-            if range.end <= range.start {
-                continue;
-            }
-
-            let mask = reveal_mask(parent_mask, bounds.origin.x, line_width, range);
-
-            window.with_content_mask(Some(mask), |window| {
-                LineLayout::paint(
-                    &self.line,
-                    bounds.origin,
-                    layout.line_height,
-                    TextAlign::Left,
-                    None,
-                    std::slice::from_ref(&layout.foreground),
-                    window,
-                    cx,
-                )
-                .ok();
-            });
-        }
-    }
-}
-
-fn reveal_mask(
-    parent: ContentMask<Pixels>,
-    origin_x: Pixels,
-    line_width: Pixels,
-    range: RevealRange,
-) -> ContentMask<Pixels> {
-    let left = match range.start <= px(0.) {
-        true => parent.bounds.left(),
-        false => origin_x + range.start,
-    };
-    let right = match range.end >= line_width {
-        true => parent.bounds.right(),
-        false => origin_x + range.end,
-    };
-    ContentMask {
-        bounds: Bounds::from_corners(
-            point(left, parent.bounds.top()),
-            point(right, parent.bounds.bottom()),
-        ),
-    }
-}
-
-#[derive(Clone, Copy, Debug, PartialEq)]
-struct RevealRange {
-    start: Pixels,
-    end: Pixels,
-}
-
-fn revealed(
-    plan: &Wrapped,
-    units: Range<usize>,
-    windows: &[(std::time::Duration, std::time::Duration)],
-    position: std::time::Duration,
-) -> Vec<RevealRange> {
-    let mut ranges: Vec<(Pixels, Pixels)> = Vec::new();
-    let mut offset = px(0.);
-    for index in units {
-        let unit = &plan.units[index];
-        for part in &unit.parts {
-            let Some(&(start, end)) = windows.get(part.word) else {
-                continue;
-            };
-
-            let share = progress_between(start, end, position);
-            if share > 0. {
-                let whole = plan
-                    .word_widths
-                    .get(part.word)
-                    .copied()
-                    .unwrap_or(part.width);
-                let progress = match part.width > px(0.) {
-                    true => ((whole * share - part.before) / part.width).clamp(0., 1.),
-                    false => 0.,
-                };
-                let start = offset + part.offset;
-                let end = start + part.width * progress;
-                if end <= start {
-                    continue;
-                }
-                match ranges.last_mut() {
-                    Some((_, previous_end)) if start <= *previous_end => {
-                        *previous_end = (*previous_end).max(end);
-                    }
-                    _ => ranges.push((start, end)),
-                }
-            }
-        }
-        offset += unit.width;
-    }
-
-    ranges
-        .into_iter()
-        .map(|(start, end)| RevealRange { start, end })
-        .collect()
-}
-
-struct SecondaryLaneLook<'a> {
-    dimming: Option<u64>,
-    voice: Voice,
-    sung: Sung,
-    plan: Option<&'a Wrapped>,
-}
-
-fn secondary_lyrics_lane(
-    lane: &music::LyricsLane,
-    line_active: bool,
-    line_passed: bool,
-    position: std::time::Duration,
-    look: SecondaryLaneLook<'_>,
-) -> gpui::AnyElement {
-    let SecondaryLaneLook {
-        dimming,
-        voice,
-        sung,
-        plan,
-    } = look;
-    let theme = &sung.theme;
-    let passed = line_passed || lane.sung_end().is_some_and(|end| position >= end);
-    let shade = |singing: bool| {
-        let active =
-            singing && position >= lane.start && lane.sung_end().is_none_or(|end| position < end);
-        let karaoke =
-            secondary_karaoke_visible(lane, singing, position) && sung.karaoke && lane.worded();
-
-        match (active, passed, karaoke) {
-            (_, _, true) => theme.muted_foreground,
-            (true, _, false) => theme.foreground,
-            (false, true, false) => theme.muted_foreground.opacity(PAST),
-            (false, false, false) => theme.muted_foreground.opacity(AHEAD),
-        }
-    };
-    let tint = shade(line_active);
-    let size = sung.lane;
-    let karaoke_capable = sung.karaoke && lane.worded();
-    let lyrics =
-        div()
-            .text_size(size)
-            .map(|this| match (karaoke_capable, lane.words.as_ref(), plan) {
-                (true, Some(words), Some(plan)) => {
-                    this.child(karaoke_lane(plan, words, position, voice, sung))
-                }
-                _ => this.child(SharedString::from(lane.text.clone())),
-            });
-    let held = shade(true);
-    let lyrics = match dimming {
-        Some(departure) => lyrics
-            .motion(("lane-dim", departure as usize), Motion::Quick, {
-                move |this, t| this.text_color(mix(held, tint, t))
-            })
-            .into_any_element(),
-        None => lyrics.text_color(tint).into_any_element(),
-    };
-    div()
-        .flex()
-        .flex_col()
-        .when(!voice.lead(), |this| this.items_end().text_right())
-        .child(lyrics)
-        .when_some(
-            selected_romanization(&lane.romanized, sung.scripts),
-            |this, text| this.child(romanized_lyrics_lane(text, size, theme)),
-        )
-        .into_any_element()
-}
-
-fn secondary_lane_started(lane: &music::LyricsLane, position: std::time::Duration) -> bool {
-    position >= lane.start
-}
-
-fn secondary_karaoke_visible(
-    lane: &music::LyricsLane,
-    line_active: bool,
-    position: std::time::Duration,
-) -> bool {
-    line_active && secondary_lane_started(lane, position)
-}
-
-fn selected_romanization(
-    romanized: &Option<music::RomanizedText>,
-    scripts: Option<RomanizationScripts>,
-) -> Option<String> {
-    let romanized = romanized.as_ref()?;
-    scripts?
-        .contains(romanized.writing_system)
-        .then(|| romanized.text.clone())
-}
-
-fn romanized_lyrics_lane(text: String, size: Pixels, theme: &ui::Theme) -> Div {
-    div()
-        .text_size(size)
-        .text_color(theme.muted_foreground)
-        .child(SharedString::from(text))
-}
-
-fn karaoke_window(
-    words: &[music::LyricsWord],
-    index: usize,
-) -> (std::time::Duration, std::time::Duration) {
-    let word = &words[index];
-    let start = word.start;
-    let end = match word.end > start {
-        true => word.end,
-        false => words
-            .get(index + 1)
-            .map(|next| next.start.max(start))
-            .filter(|end| *end > start)
-            .unwrap_or(start + SWEEP_LEAST),
-    };
-    (start, end)
-}
-
 /// How far along a transition started at this moment is, asking for frames while
 /// it runs.
 fn ramp(at: std::time::Instant, window: &mut Window) -> f32 {
@@ -2267,308 +1856,13 @@ fn ramp(at: std::time::Instant, window: &mut Window) -> f32 {
     ease_out_expo(progress)
 }
 
-/// Scales one line of text without touching the space between it and the next.
-fn lifted(row: Div, sung: Sung) -> Div {
-    match sung.lift == 1. {
-        true => row,
-        false => row.layer_scale(sung.lift).layer_scale_origin(sung.from),
-    }
-}
-
 fn active_verse_size(verse: Pixels) -> Pixels {
     verse + ACTIVE_VERSE_GROWTH
-}
-
-#[derive(Clone)]
-struct TimingPart {
-    word: usize,
-    offset: Pixels,
-    before: Pixels,
-    width: Pixels,
-}
-
-#[derive(Clone)]
-struct VisualUnit {
-    range: Range<usize>,
-    width: Pixels,
-    parts: Vec<TimingPart>,
-}
-
-// Original text remains source of truth; units and rows only hold byte ranges into it.
-#[derive(Clone)]
-struct Wrapped {
-    units: Vec<VisualUnit>,
-    rows: Vec<Range<usize>>,
-    word_widths: Vec<Pixels>,
-    text: Vec<SharedString>,
-    shapes: Vec<ShapedLine>,
-}
-
-fn lyrics_wrap_rows(
-    line: &str,
-    words: Option<&[music::LyricsWord]>,
-    font_size: Pixels,
-    width: Pixels,
-    window: &mut Window,
-) -> Option<Wrapped> {
-    (width > px(0.)).then(|| lyrics_plan(line, words, font_size, Some(width), window))
-}
-
-fn lyrics_plan(
-    line: &str,
-    words: Option<&[music::LyricsWord]>,
-    font_size: Pixels,
-    width: Option<Pixels>,
-    window: &mut Window,
-) -> Wrapped {
-    let mut style = window.text_style();
-    style.font_weight = FontWeight::SEMIBOLD;
-    let source = SharedString::from(line.to_owned());
-    let run = style.to_run(source.len());
-    let shaped = window
-        .text_system()
-        .shape_line(source.clone(), font_size, &[run], None);
-    let timing = words
-        .filter(|words| !words.is_empty())
-        .map(|words| timing_spans(line, words))
-        .unwrap_or_default();
-    let normal = normal_break_ranges(line);
-    let ranges = match width {
-        Some(width) => emergency_ranges(line, normal, width, |index| shaped.x_for_index(index)),
-        None => normal,
-    };
-    let measure = |index| shaped.x_for_index(index);
-    let (units, word_widths) = measured_units(ranges, &timing, &measure);
-    let rows = width
-        .map(|width| wrap_unit_widths(&units, width))
-        .unwrap_or_default();
-    let text = rows
-        .iter()
-        .map(|row| {
-            let start = units[row.start].range.start;
-            let end = units[row.end - 1].range.end;
-            SharedString::from(source.as_ref()[start..end].to_owned())
-        })
-        .collect::<Vec<_>>();
-    let shapes = match (words.is_some(), rows.is_empty()) {
-        (false, _) => Vec::new(),
-        (true, true) => units
-            .iter()
-            .map(|unit| shaped_range(&shaped, unit.range.clone()))
-            .collect(),
-        (true, false) => rows
-            .iter()
-            .map(|row| {
-                let start = units[row.start].range.start;
-                let end = units[row.end - 1].range.end;
-                shaped_range(&shaped, start..end)
-            })
-            .collect(),
-    };
-
-    Wrapped {
-        units,
-        rows,
-        word_widths,
-        text,
-        shapes,
-    }
-}
-
-fn shaped_range(line: &ShapedLine, range: Range<usize>) -> ShapedLine {
-    let (_, suffix) = line.split_at(range.start);
-    let (slice, _) = suffix.split_at(range.end - range.start);
-    slice
-}
-
-fn measured_units(
-    ranges: Vec<Range<usize>>,
-    timing: &[Range<usize>],
-    x_for_index: &impl Fn(usize) -> Pixels,
-) -> (Vec<VisualUnit>, Vec<Pixels>) {
-    let word_widths = timing
-        .iter()
-        .map(|span| x_for_index(span.end) - x_for_index(span.start))
-        .collect::<Vec<_>>();
-    let units = ranges
-        .into_iter()
-        .map(|range| {
-            let start_x = x_for_index(range.start);
-            let width = x_for_index(range.end) - start_x;
-            let parts = timing
-                .iter()
-                .enumerate()
-                .filter_map(|(word, span)| {
-                    let start = range.start.max(span.start);
-                    let end = range.end.min(span.end);
-                    (start < end).then(|| TimingPart {
-                        word,
-                        offset: x_for_index(start) - start_x,
-                        before: x_for_index(start) - x_for_index(span.start),
-                        width: x_for_index(end) - x_for_index(start),
-                    })
-                })
-                .collect();
-            VisualUnit {
-                range,
-                width,
-                parts,
-            }
-        })
-        .collect::<Vec<_>>();
-
-    (units, word_widths)
-}
-
-// Provider spans own timing only. They never supply line-break semantics.
-fn timing_spans(line: &str, words: &[music::LyricsWord]) -> Vec<Range<usize>> {
-    let mut starts = Vec::with_capacity(words.len());
-    let mut cursor = 0;
-    for word in words {
-        if word.text.is_empty() {
-            return approximate_timing_spans(line, words);
-        }
-        let Some(remainder) = line.get(cursor..) else {
-            return approximate_timing_spans(line, words);
-        };
-        let Some(relative) = remainder.find(&word.text) else {
-            return approximate_timing_spans(line, words);
-        };
-        let start = cursor + relative;
-        starts.push(start);
-        cursor = start + word.text.len();
-    }
-
-    starts
-        .iter()
-        .enumerate()
-        .map(|(index, start)| {
-            let start = if index == 0 { 0 } else { *start };
-            let end = starts.get(index + 1).copied().unwrap_or(line.len());
-            start..end
-        })
-        .collect()
-}
-
-// Mismatched provider text is rare after normalization. Keep original bytes and distribute
-// ownership over character boundaries so rendering stays lossless and safe.
-fn approximate_timing_spans(line: &str, words: &[music::LyricsWord]) -> Vec<Range<usize>> {
-    if words.is_empty() {
-        return Vec::new();
-    }
-    let boundaries = line
-        .char_indices()
-        .map(|(at, _)| at)
-        .chain(std::iter::once(line.len()))
-        .collect::<Vec<_>>();
-    let characters = boundaries.len() - 1;
-    let weights = words
-        .iter()
-        .map(|word| word.text.chars().count())
-        .collect::<Vec<_>>();
-    let total = weights.iter().sum::<usize>();
-    let mut consumed = 0usize;
-    let mut start = 0;
-    words
-        .iter()
-        .enumerate()
-        .map(|(index, _)| {
-            consumed += weights[index];
-            let target = if index + 1 == words.len() {
-                characters
-            } else {
-                consumed
-                    .saturating_mul(characters)
-                    .checked_div(total)
-                    .unwrap_or((index + 1) * characters / words.len())
-            };
-            let end = boundaries[target.min(characters)].max(start);
-            let span = start..end;
-            start = end;
-            span
-        })
-        .collect()
-}
-
-fn normal_break_ranges(line: &str) -> Vec<Range<usize>> {
-    // ICU4X applies UAX #14/Kinsoku rules and dictionary data for Thai and the
-    // other complex scripts. Timing units never participate in this pass.
-    let mut offsets = LineSegmenter::new_dictionary(LineBreakOptions::default())
-        .segment_str(line)
-        .filter(|offset| line.is_char_boundary(*offset))
-        .collect::<Vec<_>>();
-    if offsets.first().copied() != Some(0) {
-        offsets.insert(0, 0);
-    }
-    if offsets.last().copied() != Some(line.len()) {
-        offsets.push(line.len());
-    }
-    offsets.dedup();
-    offsets
-        .windows(2)
-        .filter_map(|pair| (pair[0] < pair[1]).then_some(pair[0]..pair[1]))
-        .collect()
-}
-
-// Only an individually oversized normal segment may enter grapheme fallback.
-fn emergency_ranges(
-    line: &str,
-    normal: Vec<Range<usize>>,
-    width: Pixels,
-    x_for_index: impl Fn(usize) -> Pixels,
-) -> Vec<Range<usize>> {
-    let mut ranges = Vec::new();
-    for range in normal {
-        if x_for_index(range.end) - x_for_index(range.start) <= width {
-            ranges.push(range);
-            continue;
-        }
-        let text = &line[range.clone()];
-        ranges.extend(text.grapheme_indices(true).map(|(at, cluster)| {
-            let start = range.start + at;
-            start..start + cluster.len()
-        }));
-    }
-    ranges
-}
-
-fn wrap_unit_widths(units: &[VisualUnit], width: Pixels) -> Vec<Range<usize>> {
-    let mut rows = Vec::new();
-    let mut start = 0;
-    let mut used = px(0.);
-
-    for (index, unit) in units.iter().enumerate() {
-        if index > start && used + unit.width > width {
-            rows.push(start..index);
-            start = index;
-            used = px(0.);
-        }
-        used += unit.width;
-    }
-    if start < units.len() {
-        rows.push(start..units.len());
-    }
-    rows
 }
 
 fn anchored_lyrics_offset(view: Pixels, item: Pixels, height: Pixels, reach: Pixels) -> Pixels {
     let delta = view - item + height * PIN;
     delta.clamp(-reach, px(0.))
-}
-
-fn progress_between(
-    start: std::time::Duration,
-    end: std::time::Duration,
-    position: std::time::Duration,
-) -> f32 {
-    if position <= start {
-        return 0.;
-    }
-    if position >= end {
-        return 1.;
-    }
-    let span = (end - start).as_secs_f32();
-    ((position - start).as_secs_f32() / span).clamp(0., 1.)
 }
 
 fn instrumental_gap_before(lines: &[music::LyricsLine], index: usize) -> std::time::Duration {
@@ -2619,58 +1913,6 @@ fn sung_line(lines: &[music::LyricsLine], position: std::time::Duration) -> Opti
         Some(_) => None,
         None => music::lyrics::active(lines, position),
     }
-}
-
-fn lanes_room(
-    lanes: &[music::LyricsLane],
-    scripts: Option<RomanizationScripts>,
-    size: Pixels,
-    leading: Pixels,
-    width: Pixels,
-    window: &mut Window,
-    plans: Option<&HashMap<usize, Wrapped>>,
-) -> Pixels {
-    let rows = lanes
-        .iter()
-        .enumerate()
-        .map(|(lane_index, lane)| {
-            let spoken = plans.and_then(|plans| plans.get(&lane_index)).map_or_else(
-                || wrapped_rows(&lane.text, size, width, window),
-                |plan| plan.rows.len().max(1),
-            );
-            let romanized = selected_romanization(&lane.romanized, scripts)
-                .map_or(0, |text| wrapped_rows(&text, size, width, window));
-            spoken + romanized
-        })
-        .sum::<usize>();
-
-    // a lane inherits the line height of the verse, not its own text size
-    let gaps = window.rem_size() * LANE_GAP_REM * lanes.len().saturating_sub(1) as f32;
-
-    leading * (rows as f32 + LANE_SLACK) + gaps
-}
-
-fn plain_lyrics_rows(
-    text: &str,
-    size: Pixels,
-    width: Pixels,
-    window: &mut Window,
-) -> Vec<SharedString> {
-    let mut rows = Vec::new();
-    for line in text.split('\n') {
-        if line.is_empty() {
-            rows.push(SharedString::from(""));
-        } else if let Some(plan) = lyrics_wrap_rows(line, None, size, width, window) {
-            rows.extend(plan.text);
-        } else {
-            rows.push(SharedString::from(line.to_owned()));
-        }
-    }
-    rows
-}
-
-fn wrapped_rows(text: &str, size: Pixels, width: Pixels, window: &mut Window) -> usize {
-    lyrics_wrap_rows(text, None, size, width, window).map_or(1, |wrapped| wrapped.rows.len().max(1))
 }
 
 // culling needs layout
@@ -2762,43 +2004,6 @@ fn viewport_along(scroll: &ScrollHandle, row: usize, view: Bounds<Pixels>, downw
     }
 }
 
-fn line_has_passed(line: &music::LyricsLine, position: std::time::Duration) -> bool {
-    line.sung_end().is_some_and(|end| position >= end)
-}
-
-fn primary_karaoke_visible(
-    line: &music::LyricsLine,
-    line_active: bool,
-    position: std::time::Duration,
-) -> bool {
-    line_active
-        || (position >= line.start
-            && line
-                .sung_end()
-                .is_some_and(|end| position < end + Motion::Control.span()))
-}
-
-fn primary_karaoke_fade(
-    line: &music::LyricsLine,
-    line_active: bool,
-    position: std::time::Duration,
-) -> f32 {
-    if line_active {
-        return 0.;
-    }
-    line.sung_end().map_or(0., |end| {
-        progress_between(end, end + Motion::Control.span(), position)
-    })
-}
-
-fn background_line_singing(
-    line: &music::LyricsLine,
-    line_active: bool,
-    position: std::time::Duration,
-) -> bool {
-    !line_active && position >= line.start && line.sung_end().is_some_and(|end| position < end)
-}
-
 fn instrumental_row(progress: f32, past: bool, verse: Pixels, theme: &ui::Theme) -> Div {
     let note_size = verse * 1.;
     div()
@@ -2839,83 +2044,16 @@ fn wordless(key: &'static str, icon: &'static str) -> gpui::AnyElement {
 
 #[cfg(test)]
 mod tests {
-    use std::{ops::Range, time::Duration};
+    use std::time::Duration;
 
-    use unicode_segmentation::UnicodeSegmentation;
-
-    use music::{LyricsLane, LyricsLine, LyricsWord, Voice};
+    use gpui::px;
+    use music::{LyricsLine, Voice};
     use ui::Springs;
 
     use super::{
-        QueuePosition, RevealRange, Sections, Slot, TimingPart, VisualUnit, Wrapped,
-        active_lyrics_row, anchored_lyrics_offset, background_line_singing, emergency_ranges,
-        karaoke_window, lag_spring, line_has_passed, line_row, lyric_row_count, measured_units,
-        normal_break_ranges, primary_karaoke_fade, primary_karaoke_visible, progress_between,
-        reveal_mask, revealed, secondary_karaoke_visible, timing_spans, wrap_unit_widths,
+        QueuePosition, Sections, Slot, active_lyrics_row, anchored_lyrics_offset, lag_spring,
+        line_row, lyric_row_count,
     };
-    use gpui::{Bounds, ContentMask, Pixels, SharedString, point, px};
-    use ui::Motion;
-
-    fn test_units(widths: &[Pixels]) -> Vec<VisualUnit> {
-        widths
-            .iter()
-            .map(|&width| VisualUnit {
-                range: 0..0,
-                width,
-                parts: Vec::new(),
-            })
-            .collect()
-    }
-
-    fn segments(text: &str) -> Vec<&str> {
-        normal_break_ranges(text)
-            .into_iter()
-            .map(|range| &text[range])
-            .collect()
-    }
-
-    fn timed_words(parts: &[&str]) -> Vec<LyricsWord> {
-        parts
-            .iter()
-            .enumerate()
-            .map(|(index, text)| LyricsWord {
-                start: Duration::from_millis(index as u64 * 100),
-                end: Duration::from_millis(index as u64 * 100 + 100),
-                text: (*text).to_owned(),
-            })
-            .collect()
-    }
-
-    fn reveal_plan() -> Wrapped {
-        Wrapped {
-            units: vec![
-                VisualUnit {
-                    range: 0..2,
-                    width: px(10.),
-                    parts: vec![TimingPart {
-                        word: 0,
-                        offset: px(0.),
-                        before: px(0.),
-                        width: px(10.),
-                    }],
-                },
-                VisualUnit {
-                    range: 3..5,
-                    width: px(10.),
-                    parts: vec![TimingPart {
-                        word: 1,
-                        offset: px(0.),
-                        before: px(0.),
-                        width: px(10.),
-                    }],
-                },
-            ],
-            rows: std::iter::once(0..2).collect(),
-            word_widths: vec![px(10.), px(10.)],
-            text: vec![SharedString::from("AA BB")],
-            shapes: Vec::new(),
-        }
-    }
 
     fn slots(sections: Sections) -> Vec<Slot> {
         (0..sections.len()).map(|i| sections.slot(i)).collect()
@@ -3053,38 +2191,6 @@ mod tests {
     }
 
     #[test]
-    fn word_timing_exposes_a_pause_hidden_by_the_line_end() {
-        let lines = [
-            LyricsLine {
-                start: Duration::from_secs(2),
-                end: Some(Duration::from_secs(12)),
-                text: "first".to_owned(),
-                romanized: None,
-                words: Some(vec![LyricsWord {
-                    start: Duration::from_secs(2),
-                    end: Duration::from_secs(5),
-                    text: "first".to_owned(),
-                }]),
-                secondary: Vec::new(),
-                voice: Voice::Lead,
-            },
-            LyricsLine {
-                start: Duration::from_secs(12),
-                end: Some(Duration::from_secs(15)),
-                text: "second".to_owned(),
-                romanized: None,
-                words: None,
-                secondary: Vec::new(),
-                voice: Voice::Lead,
-            },
-        ];
-
-        assert_eq!(lyric_row_count(&lines), 3);
-        assert_eq!(active_lyrics_row(&lines, Duration::from_secs(8)), Some(1));
-        assert!(line_has_passed(&lines[0], Duration::from_secs(8)));
-    }
-
-    #[test]
     fn lyrics_follow_uses_unscrolled_item_bounds() {
         let offset = anchored_lyrics_offset(px(0.), px(200.), px(100.), px(500.));
 
@@ -3096,95 +2202,6 @@ mod tests {
         let offset = anchored_lyrics_offset(px(0.25), px(200.125), px(100.5), px(500.));
 
         assert!((offset.as_f32() - -169.725).abs() < 0.001);
-    }
-
-    #[test]
-    fn timing_spans_keep_original_spacing() {
-        let text = "I said oooh I'm drowning in the night";
-        let words = ["I", "said", "oooh", "I'm", "drowning", "in", "the", "night"]
-            .into_iter()
-            .enumerate()
-            .map(|(index, text)| LyricsWord {
-                start: Duration::from_millis(index as u64 * 100),
-                end: Duration::from_millis(index as u64 * 100 + 100),
-                text: text.to_owned(),
-            })
-            .collect::<Vec<_>>();
-
-        let spans = timing_spans(text, &words);
-        assert_eq!(
-            spans
-                .iter()
-                .map(|range| &text[range.clone()])
-                .collect::<Vec<_>>(),
-            [
-                "I ",
-                "said ",
-                "oooh ",
-                "I'm ",
-                "drowning ",
-                "in ",
-                "the ",
-                "night"
-            ]
-        );
-    }
-
-    #[test]
-    fn normal_breaking_keeps_original_text() {
-        let text = "Ладони полны слёзок, но время";
-        let ranges = normal_break_ranges(text);
-
-        assert_eq!(ranges.first().map(|range| range.start), Some(0));
-        assert_eq!(ranges.last().map(|range| range.end), Some(text.len()));
-        assert_eq!(
-            ranges
-                .iter()
-                .map(|range| &text[range.clone()])
-                .collect::<Vec<_>>(),
-            ["Ладони ", "полны ", "слёзок, ", "но ", "время"]
-        );
-    }
-
-    #[test]
-    fn lyrics_wrap_only_at_normal_unit_boundaries() {
-        let units = test_units(&[px(60.), px(30.), px(20.)]);
-        let rows = wrap_unit_widths(&units, px(80.));
-
-        assert_eq!(rows, [0..1, 1..3]);
-    }
-
-    #[test]
-    fn lyrics_keep_an_oversized_unit_on_its_own_row() {
-        let rows = wrap_unit_widths(&test_units(&[px(120.), px(30.), px(30.)]), px(80.));
-
-        assert_eq!(rows, [0..1, 1..3]);
-    }
-
-    #[test]
-    fn timing_parts_can_cross_a_wrapped_row() {
-        let timing = [0..2, 2..5, 5..6];
-        let (units, widths) = measured_units(vec![0..3, 3..6], &timing, &|index| px(index as f32));
-
-        assert_eq!(widths, [px(2.), px(3.), px(1.)]);
-        assert_eq!(
-            units[0]
-                .parts
-                .iter()
-                .map(|part| part.word)
-                .collect::<Vec<_>>(),
-            [0, 1]
-        );
-        assert_eq!(
-            units[1]
-                .parts
-                .iter()
-                .map(|part| part.word)
-                .collect::<Vec<_>>(),
-            [1, 2]
-        );
-        assert_eq!(units[1].parts[0].before, px(1.));
-        assert_eq!(wrap_unit_widths(&units, px(3.)), [0..1, 1..2]);
     }
 
     #[test]
@@ -3208,583 +2225,5 @@ mod tests {
         assert_eq!(Springs::LYRICS_ROW.stiffness, 210.);
         assert_eq!(Springs::LYRICS_ROW.damping, 22.);
         assert_eq!(Springs::LYRICS_ROW.mass, 1.);
-    }
-
-    #[test]
-    fn a_late_first_word_starts_at_provider_time() {
-        let words = vec![
-            LyricsWord {
-                start: Duration::from_millis(1500),
-                end: Duration::from_millis(1900),
-                text: "first".to_owned(),
-            },
-            LyricsWord {
-                start: Duration::from_millis(2000),
-                end: Duration::from_millis(2400),
-                text: "second".to_owned(),
-            },
-        ];
-
-        assert_eq!(
-            karaoke_window(&words, 0),
-            (Duration::from_millis(1500), Duration::from_millis(1900))
-        );
-        let plan = reveal_plan();
-        let windows = [(Duration::from_millis(1500), Duration::from_millis(1900))];
-        assert!(revealed(&plan, 0..1, &windows, Duration::from_millis(1499)).is_empty());
-    }
-
-    #[test]
-    fn overlapping_word_windows_keep_provider_ends() {
-        let words = vec![
-            LyricsWord {
-                start: Duration::from_millis(1000),
-                end: Duration::from_millis(2000),
-                text: "A".to_owned(),
-            },
-            LyricsWord {
-                start: Duration::from_millis(1500),
-                end: Duration::from_millis(2500),
-                text: "B".to_owned(),
-            },
-        ];
-
-        assert_eq!(
-            karaoke_window(&words, 0),
-            (Duration::from_millis(1000), Duration::from_millis(2000))
-        );
-        assert_eq!(
-            karaoke_window(&words, 1),
-            (Duration::from_millis(1500), Duration::from_millis(2500))
-        );
-    }
-
-    #[test]
-    fn zero_duration_word_uses_only_limited_fallback() {
-        let words = vec![
-            LyricsWord {
-                start: Duration::from_millis(1000),
-                end: Duration::from_millis(1000),
-                text: "A".to_owned(),
-            },
-            LyricsWord {
-                start: Duration::from_millis(1500),
-                end: Duration::from_millis(2000),
-                text: "B".to_owned(),
-            },
-        ];
-
-        assert_eq!(
-            karaoke_window(&words, 0),
-            (Duration::from_millis(1000), Duration::from_millis(1500))
-        );
-        assert_eq!(
-            karaoke_window(&words[..1], 0),
-            (Duration::from_millis(1000), Duration::from_millis(1180))
-        );
-    }
-
-    #[test]
-    fn positive_duration_progress_is_linear() {
-        let start = Duration::from_millis(1000);
-        let end = Duration::from_millis(2000);
-
-        assert_eq!(
-            progress_between(start, end, Duration::from_millis(999)),
-            0.0
-        );
-        assert_eq!(
-            progress_between(start, end, Duration::from_millis(1000)),
-            0.0
-        );
-        assert_eq!(
-            progress_between(start, end, Duration::from_millis(1250)),
-            0.25
-        );
-        assert_eq!(
-            progress_between(start, end, Duration::from_millis(1500)),
-            0.5
-        );
-        assert_eq!(
-            progress_between(start, end, Duration::from_millis(1750)),
-            0.75
-        );
-        assert_eq!(
-            progress_between(start, end, Duration::from_millis(2000)),
-            1.0
-        );
-        assert_eq!(
-            progress_between(start, end, Duration::from_millis(2500)),
-            1.0
-        );
-    }
-
-    #[test]
-    fn sequential_words_join_continuously_at_boundary() {
-        let a_start = Duration::from_millis(1000);
-        let a_end = Duration::from_millis(1500);
-        let b_start = Duration::from_millis(1500);
-        let b_end = Duration::from_millis(2000);
-
-        let boundary = Duration::from_millis(1500);
-        assert_eq!(progress_between(a_start, a_end, boundary), 1.0);
-        assert_eq!(progress_between(b_start, b_end, boundary), 0.0);
-    }
-
-    #[test]
-    fn reveal_mask_geometry_rules() {
-        let parent = ContentMask {
-            bounds: Bounds::from_corners(point(px(10.), px(20.)), point(px(200.), px(100.))),
-        };
-        let origin_x = px(100.);
-        let line_width = px(50.);
-
-        let interior = reveal_mask(
-            parent,
-            origin_x,
-            line_width,
-            RevealRange {
-                start: px(20.),
-                end: px(40.),
-            },
-        );
-        assert_eq!(interior.bounds.left(), px(120.));
-        assert_eq!(interior.bounds.right(), px(140.));
-        assert_eq!(interior.bounds.top(), px(20.));
-        assert_eq!(interior.bounds.bottom(), px(100.));
-
-        let starts_at_edge = reveal_mask(
-            parent,
-            origin_x,
-            line_width,
-            RevealRange {
-                start: px(0.),
-                end: px(40.),
-            },
-        );
-        assert_eq!(starts_at_edge.bounds.left(), px(10.));
-        assert_eq!(starts_at_edge.bounds.right(), px(140.));
-        assert_eq!(starts_at_edge.bounds.top(), px(20.));
-        assert_eq!(starts_at_edge.bounds.bottom(), px(100.));
-
-        let terminal_complete = reveal_mask(
-            parent,
-            origin_x,
-            line_width,
-            RevealRange {
-                start: px(30.),
-                end: px(50.),
-            },
-        );
-        assert_eq!(terminal_complete.bounds.left(), px(130.));
-        assert_eq!(terminal_complete.bounds.right(), px(200.));
-        assert_eq!(terminal_complete.bounds.top(), px(20.));
-        assert_eq!(terminal_complete.bounds.bottom(), px(100.));
-
-        let row_complete = reveal_mask(
-            parent,
-            origin_x,
-            line_width,
-            RevealRange {
-                start: px(0.),
-                end: px(50.),
-            },
-        );
-        assert_eq!(row_complete.bounds.left(), px(10.));
-        assert_eq!(row_complete.bounds.right(), px(200.));
-        assert_eq!(row_complete.bounds.top(), px(20.));
-        assert_eq!(row_complete.bounds.bottom(), px(100.));
-    }
-
-    #[test]
-    fn overlapping_reveal_ranges_do_not_fill_the_gap() {
-        let plan = reveal_plan();
-        let windows = [
-            (Duration::from_millis(1000), Duration::from_millis(2000)),
-            (Duration::from_millis(1500), Duration::from_millis(2500)),
-        ];
-        let reveal = revealed(
-            &plan,
-            0..plan.units.len(),
-            &windows,
-            Duration::from_millis(1750),
-        );
-
-        assert_eq!(reveal.len(), 2);
-        assert_eq!(reveal[0].start, px(0.));
-        assert_eq!(reveal[0].end, px(7.5));
-        assert_eq!(reveal[1].start, px(10.));
-        assert_eq!(reveal[1].end, px(12.5));
-    }
-
-    #[test]
-    fn sequential_reveal_ranges_still_join_continuously() {
-        let plan = reveal_plan();
-        let windows = [
-            (Duration::from_millis(1000), Duration::from_millis(1500)),
-            (Duration::from_millis(1500), Duration::from_millis(2000)),
-        ];
-        let reveal = revealed(
-            &plan,
-            0..plan.units.len(),
-            &windows,
-            Duration::from_millis(1750),
-        );
-
-        assert_eq!(reveal.len(), 1);
-        assert_eq!(reveal[0].start, px(0.));
-        assert_eq!(reveal[0].end, px(15.));
-    }
-
-    #[test]
-    fn a_finished_background_lane_stays_sung_until_its_line_departs() {
-        let lane = LyricsLane {
-            start: Duration::from_secs(2),
-            end: Some(Duration::from_secs(3)),
-            text: "(E)".to_owned(),
-            romanized: None,
-            words: Some(vec![LyricsWord {
-                start: Duration::from_secs(2),
-                end: Duration::from_secs(3),
-                text: "(E)".to_owned(),
-            }]),
-        };
-
-        assert!(!secondary_karaoke_visible(
-            &lane,
-            true,
-            Duration::from_millis(1999)
-        ));
-        assert!(secondary_karaoke_visible(
-            &lane,
-            true,
-            Duration::from_secs(2)
-        ));
-        assert!(secondary_karaoke_visible(
-            &lane,
-            true,
-            Duration::from_secs(4)
-        ));
-        assert!(!secondary_karaoke_visible(
-            &lane,
-            false,
-            Duration::from_secs(4)
-        ));
-    }
-
-    #[test]
-    fn an_overlapped_primary_line_keeps_singing_in_the_background() {
-        let line = LyricsLine {
-            start: Duration::from_secs(2),
-            end: Some(Duration::from_secs(8)),
-            text: "Wake me up inside".to_owned(),
-            romanized: None,
-            words: Some(vec![LyricsWord {
-                start: Duration::from_secs(2),
-                end: Duration::from_secs(8),
-                text: "Wake me up inside".to_owned(),
-            }]),
-            secondary: Vec::new(),
-            voice: Voice::Lead,
-        };
-
-        assert!(!primary_karaoke_visible(
-            &line,
-            false,
-            Duration::from_millis(1999)
-        ));
-        assert!(primary_karaoke_visible(
-            &line,
-            false,
-            Duration::from_secs(5)
-        ));
-        assert!(primary_karaoke_visible(
-            &line,
-            false,
-            Duration::from_secs(8)
-        ));
-        assert!(!primary_karaoke_visible(
-            &line,
-            false,
-            Duration::from_secs(8) + Motion::Base.span()
-        ));
-    }
-
-    #[test]
-    fn the_active_primary_line_keeps_its_completed_sweep_until_departure() {
-        let line = LyricsLine {
-            start: Duration::from_secs(2),
-            end: Some(Duration::from_secs(5)),
-            text: "line".to_owned(),
-            romanized: None,
-            words: Some(vec![LyricsWord {
-                start: Duration::from_secs(2),
-                end: Duration::from_secs(5),
-                text: "line".to_owned(),
-            }]),
-            secondary: Vec::new(),
-            voice: Voice::Lead,
-        };
-
-        assert!(primary_karaoke_visible(&line, true, Duration::from_secs(8)));
-    }
-
-    #[test]
-    fn a_finished_background_line_fades_from_white_to_gray() {
-        let line = LyricsLine {
-            start: Duration::from_secs(2),
-            end: Some(Duration::from_secs(8)),
-            text: "Wake me up inside".to_owned(),
-            romanized: None,
-            words: Some(vec![LyricsWord {
-                start: Duration::from_secs(2),
-                end: Duration::from_secs(8),
-                text: "Wake me up inside".to_owned(),
-            }]),
-            secondary: Vec::new(),
-            voice: Voice::Lead,
-        };
-        let fade = Motion::Control.span();
-
-        assert_eq!(
-            primary_karaoke_fade(&line, false, Duration::from_millis(7999)),
-            0.
-        );
-        assert_eq!(
-            primary_karaoke_fade(&line, false, Duration::from_secs(8) + fade / 2),
-            0.5
-        );
-        assert_eq!(
-            primary_karaoke_fade(&line, false, Duration::from_secs(8) + fade),
-            1.
-        );
-        assert_eq!(
-            primary_karaoke_fade(&line, true, Duration::from_secs(8) + fade),
-            0.
-        );
-    }
-
-    #[test]
-    fn only_a_currently_singing_background_line_gets_the_reduced_blur() {
-        let line = LyricsLine {
-            start: Duration::from_secs(2),
-            end: Some(Duration::from_secs(8)),
-            text: "Wake me up inside".to_owned(),
-            romanized: None,
-            words: Some(vec![LyricsWord {
-                start: Duration::from_secs(2),
-                end: Duration::from_secs(8),
-                text: "Wake me up inside".to_owned(),
-            }]),
-            secondary: Vec::new(),
-            voice: Voice::Lead,
-        };
-
-        assert!(!background_line_singing(
-            &line,
-            false,
-            Duration::from_millis(1999)
-        ));
-        assert!(background_line_singing(
-            &line,
-            false,
-            Duration::from_secs(5)
-        ));
-        assert!(!background_line_singing(
-            &line,
-            false,
-            Duration::from_secs(8)
-        ));
-        assert!(!background_line_singing(
-            &line,
-            true,
-            Duration::from_secs(5)
-        ));
-    }
-
-    #[test]
-    fn a_finished_line_stays_past_during_a_gap() {
-        let line = LyricsLine {
-            start: Duration::from_secs(2),
-            end: Some(Duration::from_secs(5)),
-            text: "line".to_owned(),
-            romanized: None,
-            words: None,
-            secondary: Vec::new(),
-            voice: Voice::Lead,
-        };
-
-        assert!(line_has_passed(&line, Duration::from_secs(8)));
-    }
-
-    #[test]
-    fn thai_normal_breaks_ignore_timing_granularity() {
-        let text = "เข้าใจสักที";
-        let expected = ["เข้าใจ", "สัก", "ที"];
-        let granularities = [
-            ["เข้า", "ใจ", "สั", "ก", "ที"].as_slice(),
-            ["เข้าใจ", "สัก", "ที"].as_slice(),
-            ["เข้าใจสักที"].as_slice(),
-        ];
-
-        assert_eq!(segments(text), expected);
-        for parts in granularities {
-            let words = timed_words(parts);
-            let spans = timing_spans(text, &words);
-            let (units, _) =
-                measured_units(normal_break_ranges(text), &spans, &|index| px(index as f32));
-            assert_eq!(
-                spans
-                    .iter()
-                    .map(|range| &text[range.clone()])
-                    .collect::<Vec<_>>(),
-                parts
-            );
-            assert_eq!(wrap_unit_widths(&units, px(20.)), [0..1, 1..3]);
-        }
-    }
-
-    #[test]
-    fn mismatched_timing_still_owns_the_original_line() {
-        let text = "เข้าใจสักที";
-        let words = timed_words(&["เข้าใจ", "missing"]);
-        let spans = timing_spans(text, &words);
-
-        assert_eq!(
-            spans
-                .iter()
-                .map(|range| &text[range.clone()])
-                .collect::<String>(),
-            text
-        );
-        assert!(spans.iter().all(|range| {
-            text.is_char_boundary(range.start) && text.is_char_boundary(range.end)
-        }));
-    }
-
-    #[test]
-    fn required_thai_phrases_use_dictionary_boundaries() {
-        for (text, expected) in [
-            (
-                "ฉันรักเธอมากที่สุด",
-                ["ฉัน", "รัก", "เธอ", "มาก", "ที่สุด"].as_slice(),
-            ),
-            ("กรุงเทพมหานคร", ["กรุงเทพมหานคร"].as_slice()),
-            (
-                "อยากให้เข้าใจกันสักที",
-                ["อยาก", "ให้", "เข้าใจ", "กัน", "สัก", "ที"].as_slice(),
-            ),
-            ("ประเทศไทย", ["ประเทศไทย"].as_slice()),
-        ] {
-            assert_eq!(segments(text), expected);
-        }
-    }
-
-    #[test]
-    fn provider_grapheme_splits_do_not_create_normal_breaks() {
-        for (text, parts) in [
-            ("สัก", ["ส", "ั", "ก"].as_slice()),
-            ("ที่", ["ท", "ี่"].as_slice()),
-        ] {
-            let words = timed_words(parts);
-            let spans = timing_spans(text, &words);
-            let normal = normal_break_ranges(text);
-            let normal_offsets = normal
-                .iter()
-                .flat_map(|range| [range.start, range.end])
-                .collect::<Vec<_>>();
-
-            assert_eq!(
-                spans
-                    .iter()
-                    .map(|range| &text[range.clone()])
-                    .collect::<Vec<_>>(),
-                parts
-            );
-            assert_eq!(segments(text), [text]);
-            assert!(
-                spans
-                    .iter()
-                    .skip(1)
-                    .all(|range| !normal_offsets.contains(&range.start))
-            );
-        }
-    }
-
-    #[test]
-    fn emergency_ranges_split_only_the_oversized_normal_unit() {
-        let text = "a bbbbb c";
-        let normal = normal_break_ranges(text);
-        let ranges = emergency_ranges_for_test(text, normal.clone(), px(3.));
-        let pieces = ranges
-            .iter()
-            .map(|range| &text[range.clone()])
-            .collect::<Vec<_>>();
-
-        assert_eq!(
-            normal
-                .iter()
-                .map(|range| &text[range.clone()])
-                .collect::<Vec<_>>(),
-            ["a ", "bbbbb ", "c"]
-        );
-        assert_eq!(pieces.first(), Some(&"a "));
-        assert_eq!(pieces.last(), Some(&"c"));
-        assert_eq!(pieces[1..pieces.len() - 1].concat(), "bbbbb ");
-        assert!(pieces.len() > normal.len());
-    }
-
-    #[test]
-    fn emergency_ranges_preserve_extended_graphemes() {
-        for text in ["สัก", "ที่", "มากๆ", "กรุงเทพฯ"] {
-            let normal = normal_break_ranges(text);
-            let ranges = emergency_ranges_for_test(text, normal, px(1.));
-            let expected = text
-                .grapheme_indices(true)
-                .map(|(at, cluster)| at..at + cluster.len())
-                .collect::<Vec<_>>();
-
-            assert_eq!(ranges, expected, "{text}");
-        }
-    }
-
-    #[test]
-    fn cjk_and_kinsoku_boundaries_survive() {
-        assert_eq!(segments("你好世界"), ["你", "好", "世", "界"]);
-        assert_eq!(segments("「你好」"), ["「你", "好」"]);
-
-        for text in [
-            "いゝ", "いゞ", "イヽ", "イヾ", "あっ", "アッ", "日々", "あー", "あ…", "あ、", "あ。",
-        ] {
-            assert_eq!(segments(text), [text], "{text}");
-        }
-    }
-
-    #[test]
-    fn mixed_script_breaks_reconstruct_the_source() {
-        for text in ["Sonora เพลงดีมาก", "เพลงดีมาก Sonora", "Hello世界 เพลง"]
-        {
-            let ranges = normal_break_ranges(text);
-            assert_eq!(ranges.first().map(|range| range.start), Some(0));
-            assert_eq!(ranges.last().map(|range| range.end), Some(text.len()));
-            assert_eq!(
-                ranges
-                    .iter()
-                    .map(|range| &text[range.clone()])
-                    .collect::<String>(),
-                text
-            );
-            assert!(ranges.iter().all(
-                |range| text.is_char_boundary(range.start) && text.is_char_boundary(range.end)
-            ));
-        }
-    }
-
-    fn emergency_ranges_for_test(
-        text: &str,
-        normal: Vec<Range<usize>>,
-        width: Pixels,
-    ) -> Vec<Range<usize>> {
-        emergency_ranges(text, normal, width, |index| px(index as f32))
     }
 }
